+++ capnp-rpc/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.21.0
 - Follow v0.21.0 release of other capnp crates.
+- Remove deprecated `new_promise_client()` function.
 
 ## v0.20.3
 - Remove redundant information in the `reason` field of some RPC errors.